# YJIT: Change the default --yjit-call-threshold to 30